% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bdtnp}
\alias{bdtnp}
\title{Data given from the SCTC Dream Challenges}
\format{A data frame with 3039 rows for cells and 84 columns for genes}
\usage{
bdtnp
}
\description{
Data given from the SCTC Dream Challenges
}
\keyword{datasets}
