% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lin.R
\name{build_lin}
\alias{build_lin}
\title{build_lin}
\usage{
build_lin(lin_dir)
}
\arguments{
\item{lin_dir}{directory to python code}
}
\value{
3 lists of 20, 40, 60 selected genes
}
\description{
build_lin
}
