% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dge_normalized}
\alias{dge_normalized}
\title{Data given from the SCTC Dream Challenges}
\format{A data frame with 3039 rows for cells and 84 columns for genes}
\usage{
dge_normalized
}
\description{
The normalized single-cell RNA sequencing data
}
\keyword{datasets}
