% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dge_raw}
\alias{dge_raw}
\title{Data given from the SCTC Dream Challenges}
\format{A data frame with 3039 rows for cells and 84 columns for genes}
\usage{
dge_raw
}
\description{
The raw single-cell RNA sequencing data
}
\keyword{datasets}
