% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dorsal_ectoderm}
\alias{Dorsal_ectoderm}
\title{Data given from the SCTC Dream Challenges}
\format{A set of 9 genes markers for Dorsal ectoderm}
\usage{
Dorsal_ectoderm
}
\description{
Gene determinants of Dorsal ectoderm
}
\keyword{datasets}
