% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dremi}
\alias{dremi}
\title{Data given from the SCTC Dream Challenges}
\format{A data frame with 3039 rows for cells and 84 columns for genes}
\usage{
dremi
}
\description{
The distance between genes after using DREMI by Smita et al.
}
\keyword{datasets}
