% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes84}
\alias{genes84}
\title{Data given from the SCTC Dream Challenges}
\format{84-gene set}
\usage{
genes84
}
\description{
A set of 84 genes markers provided by the Dream Challenges
}
\keyword{datasets}
