% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geometry}
\alias{geometry}
\title{Data given from the SCTC Dream Challenges}
\format{A data frame with 3039 rows for cells and 3 columns for coordinations}
\usage{
geometry
}
\description{
One half of Drosophila embryo has 3039 cells places as x, y and z for a total of 3039 embryo locations rows and a total of 3039*3 coordinates
}
\keyword{datasets}
