% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{Listbycor}
\alias{Listbycor}
\title{Listbycor}
\usage{
Listbycor(normcells = NULL, normloc = NULL, method = NULL)
}
\arguments{
\item{normcells}{gene expression profiles in single cells.}

\item{normloc}{gene expression profiles in real locations.}

\item{method}{a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}
}
\value{
10 most likely locations.
}
\description{
Choose top 10 locations according cor
}
