% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{ListbyMCC}
\alias{ListbyMCC}
\title{ListbyMCC}
\usage{
ListbyMCC(bicells84 = NULL, biloc84 = NULL, geneList = NULL)
}
\arguments{
\item{bicells84}{binary gene expression profiles in single cells.}

\item{biloc84}{binary gene expression profiles in real locations.}

\item{geneList}{driver genes.}
}
\value{
10 most likely locations.
}
\description{
Choose top 10 locations according MCC
}
