% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mesoderm}
\alias{Mesoderm}
\title{Data given from the SCTC Dream Challenges}
\format{A set of 13 genes markers for Mesoderm}
\usage{
Mesoderm
}
\description{
Gene determinants of Mesoderm
}
\keyword{datasets}
