% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Neurectoderm}
\alias{Neurectoderm}
\title{Data given from the SCTC Dream Challenges}
\format{A set of 9 genes markers for Neurectoderm}
\usage{
Neurectoderm
}
\description{
Data given from the SCTC Dream Challenges
}
\keyword{datasets}
