% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NoOfGenes}
\alias{NoOfGenes}
\title{Data given from the SCTC Dream Challenges}
\format{A total number of genes used to detect locations}
\usage{
NoOfGenes
}
\description{
Total number of genes used to detect locations
}
\keyword{datasets}
