% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{numofbins}
\alias{numofbins}
\title{numofbins}
\usage{
numofbins(biExprs84 = NULL, geneList = NULL)
}
\arguments{
\item{biExprs84}{binary gene expression profiles in single cells or locations.}

\item{geneList}{driver genes.}
}
\value{
numbers of unique bins in locations and cells.
}
\description{
numofbins
}
