% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot2DCellPositions}
\alias{plot2DCellPositions}
\title{plot2DCellPositions}
\usage{
plot2DCellPositions(geometry = NULL, locidx = NULL, xcoord, zcoord)
}
\arguments{
\item{geometry}{geometry data of locations.}

\item{locidx}{Numeric vector of predicted locations.}

\item{xcoord}{x coordinate}

\item{zcoord}{z coordinate}
}
\description{
plot2DCellPositions
}
