% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot2DgenePattern}
\alias{plot2DgenePattern}
\title{plot2DgenePattern}
\usage{
plot2DgenePattern(geometry = NULL, locidx = NULL, geneProfile = NULL,
  xcoord, zcoord)
}
\arguments{
\item{geometry}{geometry data of locations.}

\item{locidx}{Numeric vector of predicted locations.}

\item{geneProfile}{gene profile of cells}

\item{xcoord}{x coordinate}

\item{zcoord}{z coordinate}
}
\description{
plot2DgenePattern
}
