% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot3DCellPositions}
\alias{plot3DCellPositions}
\title{plot3DCellPositions}
\usage{
plot3DCellPositions(geometry = NULL, locidx = NULL)
}
\arguments{
\item{geometry}{geometry data of locations.}

\item{locidx}{Numeric vector of predicted locations.}
}
\description{
plot3DCellPositions
}
