% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot3DgenePattern}
\alias{plot3DgenePattern}
\title{plot3DgenePattern}
\usage{
plot3DgenePattern(geometry = NULL, locidx = NULL, geneProfile = NULL)
}
\arguments{
\item{geometry}{geometry data of locations.}

\item{locidx}{Numeric vector of predicted locations.}

\item{geneProfile}{gene profile of cells}
}
\description{
plot3DgenePattern
}
