% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pole_cells}
\alias{Pole_cells}
\title{Data given from the SCTC Dream Challenges}
\format{1 genes markers for Pole cells}
\usage{
Pole_cells
}
\description{
Gene determinants of Pole cells
}
\keyword{datasets}
