% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc_MAGIC_dremi}
\alias{preproc_MAGIC_dremi}
\title{preproc_MAGIC_dremi}
\usage{
preproc_MAGIC_dremi(data, gene_markers)
}
\arguments{
\item{data}{Expression data with row as genes columns as samples or cells}

\item{gene_markers}{A list of gene markers e.g. 84 genes, to get distances}
}
\value{
Distance between gene markers
}
\description{
Preprocessing expression data by MAGIC and get distance by dremi
}
