% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{QEvaluation}
\alias{QEvaluation}
\title{QEvaluation}
\usage{
QEvaluation(realLoc = NULL, predloc = NULL, geometry = NULL,
  plotDist = FALSE)
}
\arguments{
\item{realLoc}{real location.}

\item{predloc}{predicted locations.}

\item{geometry}{geometry data of locations.}

\item{plotDist}{TRUE if user wants to plot distribution of detailed results}
}
\value{
sum of average distances.
}
\description{
QEvaluation
}
