% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{selectFarGenes}
\alias{selectFarGenes}
\title{selectFarGenes}
\usage{
selectFarGenes(Eucl)
}
\arguments{
\item{Eucl}{Euclidean distance matrix between 84 genes}
}
\value{
3 lists of 20, 40, 60 selected genes
}
\description{
selectFarGenes
}
