% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{selectGenesByInfluenceWithSeeds}
\alias{selectGenesByInfluenceWithSeeds}
\title{selectGenesByInfluenceWithSeeds}
\usage{
selectGenesByInfluenceWithSeeds(seedList, dremi)
}
\arguments{
\item{seedList}{the initial seed list of genes}

\item{dremi}{distance matrix between 84 genes preprocessed by DREMI}
}
\value{
3 lists of 20, 40, 60 selected genes
}
\description{
selectGenesByInfluenceWithSeeds
}
