% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{selectGenesByMADExpression}
\alias{selectGenesByMADExpression}
\title{selectGenesByMADExpression}
\usage{
selectGenesByMADExpression(dge_normalized, Eucl)
}
\arguments{
\item{dge_normalized}{normalized single-cell matrix, rows for genes and columns for cells}

\item{Eucl}{Euclidean distance matrix between 84 genes}
}
\value{
3 lists of 20, 40, 60 selected genes
}
\description{
selectGenesByMADExpression
}
