% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{selectHighRakingGenesByInfluence}
\alias{selectHighRakingGenesByInfluence}
\title{selectHighRakingGenesByInfluence}
\usage{
selectHighRakingGenesByInfluence(dremi, Eucl)
}
\arguments{
\item{dremi}{distance matrix between 84 genes preprocessed by DREMI}

\item{Eucl}{Euclidean distance matrix between 84 genes}
}
\value{
3 lists of 20, 40, 60 selected genes
}
\description{
selectHighRakingGenesByInfluence
}
