% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Yolk_cells}
\alias{Yolk_cells}
\title{Data given from the SCTC Dream Challenges}
\format{A set of 7 genes markers for Yolk cells}
\usage{
Yolk_cells

Yolk_cells
}
\description{
Gene determinants of Yolk cells

Gene determinants of Yolk cells
}
\keyword{datasets}
